Expose `IReadOnlyPathInfoStore`
Related to zjklee/Handlebars.CSharp/issues/27
